Rename functions to lasagna-master
* Renamed functions to lasagna-master to reflect concept -> exercise switch

* fix dead links and added new editor key to config
