--- exercises/concept/functions/.meta/config.json
@@ -1,8 +1,0 @@
-{
-  "authors": [
-    {
-      "github_username": "wneumann",
-      "exercism_username": "wneumann"
-    }
-  ]
-}
