--- exercises/concept/functions/Tests/LinuxMain.swift
@@ -1,6 +1,0 @@
-import XCTest
-import functionsTests
-
-var tests = [XCTestCaseEntry]()
-tests += functionsTests.allTests()
-XCTMain(tests)
