+++ exercises/concept/lasagna-master/Tests/LasagnaMasterTests/XCTestManifests.swift
@@ -3,7 +3,7 @@ import XCTest
 #if !canImport(ObjectiveC)
   public func allTests() -> [XCTestCaseEntry] {
     return [
-      testCase(functionsTests.allTests)
+      testCase(LasagnaMasterTests.allTests)
     ]
   }
 #endif
