+++ exercises/concept/lasagna-master/Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import LasagnaMasterTests
+import XCTest
+
+var tests = [XCTestCaseEntry]()
+tests += LasagnaMasterTests.allTests()
+XCTMain(tests)
